% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pbeta.R
\name{Pbeta}
\alias{Pbeta}
\title{A function to compute a Monte Carlo estimate of the Beta cdf with integer parameters}
\usage{
Pbeta(y, alpha, beta, m)
}
\arguments{
\item{y}{the parameter of cdf}

\item{alpha, beta}{the  shape paramaters of beta distribution}

\item{m}{the number of Monte Carlo experiment times}
}
\value{
a Monte Carlo estimate of the beta cdf
}
\description{
A function to compute a Monte Carlo estimate of the Beta cdf with integer parameters
}
\examples{
\dontrun{
alpha<-3;beta<-3;m<-1e4
Pbeta1<-function(x){
 Pbeta(x,alpha,beta,m)
}
sapply((1:9)/10,Pbeta1)
pbeta1<-function(x){
 pbeta(x,alpha,beta)
}
sapply((1:9)/10,pbeta1)
}
}
