% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation1.R
\name{Simulation1}
\alias{Simulation1}
\title{Simulation model 1}
\usage{
Simulation1(Q, n, p, q, rho_x, rho_e, s2n, p0, q0, nrank)
}
\arguments{
\item{Q}{parameter to specify seeds}

\item{n, p, q}{model dimensions}

\item{rho_x}{correlation parameter in the generation of predictors}

\item{rho_e}{correlation parameter in the generation of random errors}

\item{s2n}{signal to noise ratio}

\item{p0, q0}{parameters for model sparsity}

\item{nrank}{model rank}
}
\value{
similated model and data
}
\description{
similar to the the simulation model in SRRR (Lisha Chen & Jianhua Z. Huang 2012), JASS
}
\examples{
\dontrun{
data<-Simulation1(Q=1e3+10,n=50,p=20,q=10,rho_x=0.0,rho_e=0.0,s2n=1,p0=10,q0=5,nrank=3)
}
}
