% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srrr_fit.R
\name{srrr.fit}
\alias{srrr.fit}
\title{Fitting Sparse Reduced-Rank Regression}
\usage{
srrr.fit(Y, X, lam, nrank, epsilon, maxit)
}
\arguments{
\item{Y}{a matrix of response (n by q)}

\item{X}{a matrix of covariate (n by p)}

\item{lam}{penalty parameter for B}

\item{nrank}{an integer specifying the desired rank}

\item{epsilon}{convergence tolerance}

\item{maxit}{maximum number of iterations}
}
\value{
a list of model parameters
}
\description{
Given a response matrix and a covariate matrix, fits sparse reduced-Rank
regression for a specified rank.
}
\examples{
\dontrun{
data<-Simulation1(Q=1e3+10,n=50,p=20,q=10,rho_x=0.0,rho_e=0.0,s2n=1,p0=10,q0=5,nrank=3)
Result<-srrr.fit(Y=data$Y,X=data$X,lam=1,nrank=3,epsilon=1e-3,maxit=1e4)
}
}
